requirements.txt
openai==1.51.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.3
python-dateutil==2.8.2
matplotlib==3.8.2